// lib/types.ts
export interface Release {
  slug: string;
  title: string;
  artist: string;
  artistSlug: string;
  album: string;
  releaseType: string;
  trackNumber: string;
  cover: string;
  lyrics: string;
  audioUrl: string;
  releaseDate: string;
  language: string;
  genre: string;
  description: string;
  Spotify: string;
  "YouTube Music": string;
  "Apple Music": string;
  status: string;
}